Initial plan request builder impl
